requirements.txt
Django
djangorestframework
python-dotenv
psycopg2-binary
django-cors-headers
dj-database-url
pytest
pytest-django